requirements.txt
langchain==0.1.12
langchain-aws==0.1.1
PyGithub==2.4.0
boto3>=1.34.0
python-dotenv==1.0.1
pydantic>=2.0.0
requests>=2.31.0
